requirements.txt
#streamlit==1.30.0
PyYAML==6.0.1
numpy
faiss-cpu
ollama==0.1.9
requests==2.32.3
streamlit==1.35.0
sentence-transformers==2.7.0
chromadb==0.5.0
PyPDF2==3.0.1
python-docx==1.1.2
markdown==3.6
python-dotenv==1.0.1
deepseek==1.0.0
openai==1.30.1
rank_bm25==0.2.2
spacy
torch>=2.0.1,<3.0.0
torchvision>=0.15.2
transformers>=4.0.0
accelerate
bitsandbytes
huggingface_hub>=0.22.0
transformers 
accelerate 
bitsandbytes 